+++ monopyly/banking/transactions.py
@@ -139,34 +139,34 @@ class BankTransactionHandler(DatabaseHandler):
         transaction = super().update_entry(entry_id, mapping)
         return transaction
 
-    def get_matching_transactions(self, transaction_id):
+    def get_associated_transaction(self, transaction_id):
         """
-        Find all internal transactions that match this transaction.
+        Find an internal transaction that matches this transaction.
 
-        Checks all transaction databases for transactions that match the
-        given transaction.
+        Checks all transaction databases for a transaction that matches
+        the given transaction.
 
         Parameters
         ----------
         transaction_id : integer
-            The ID for the transaction that should be matched.
+            The ID for the bank transaction that should be matched.
 
         Returns
         -------
-        matching_transactions : dict
+        associated_transaction : dict
             A dictionary of transaction types and the corresponding
-            transactions that match the given transaction. The
-            transactions may be rows in either the `bank_transactions` or
+            transaction that matches the given transaction. The
+            transaction may be a row in either the `bank_transactions` or
             `credit_transactions` databases, depending on the dictionary
             key (either 'bank' or 'credit' respectively). If no matching
-            transactions are found, `None` is returned.
+            transaction is found, `None` is returned.
         """
         transaction = self.get_entry(transaction_id,
-                                     ('internal_transaction_id'))
+                                     ('internal_transaction_id',))
         internal_transaction_id = transaction['internal_transaction_id']
         if not internal_transaction_id:
             return None
-        matching_transactions = {'bank': None, 'credit': None}
+        associated_transaction = {'bank': None, 'credit': None}
         # Get matching bank transactions
         query = ("SELECT * "
                  "  FROM bank_transactions_view AS t "
@@ -174,12 +174,14 @@ class BankTransactionHandler(DatabaseHandler):
                  "          ON a.id = t.account_id "
                  "       INNER JOIN banks AS b "
                  "          ON b.id = a.bank_id "
+                 "       INNER JOIN bank_account_types AS types "
+                 "          ON types.id = a.account_type_id "
                  " WHERE b.user_id =? AND t.id != ? "
                  "       AND t.internal_transaction_id = ?")
         placeholders = (self.user_id, transaction_id, internal_transaction_id)
-        bank_transactions = self.query_entry(query, placeholders)
+        bank_transactions = self._query_entries(query, placeholders)
         if bank_transactions:
-            matching_transactions['bank'] = bank_transactions
+            associated_transaction['bank'] = bank_transactions[0]
         # Get matching credit transactions
         query = ("SELECT * "
                  "  FROM credit_transactions_view AS t "
@@ -194,9 +196,62 @@ class BankTransactionHandler(DatabaseHandler):
                  " WHERE b.user_id =? AND t.id != ? "
                  "       AND t.internal_transaction_id = ?")
         placeholders = (self.user_id, transaction_id, internal_transaction_id)
-        credit_transactions = self.query_entry(query, placeholders)
-        credit_transactions = self.query_entry(query, placeholders)
+        credit_transactions = self._query_entries(query, placeholders)
         if credit_transactions:
-            matching_transactions['credit'] = credit_transactions
-        return matching_transactions
+            associated_transaction['credit'] = credit_transactions[0]
+        return associated_transaction
+
+
+def save_transaction(form, transaction_id=None):
+    """
+    Save a banking transaction.
+
+    Saves a transaction in the database. If a transaction ID is given,
+    then the transaction is updated with the form information. Otherwise
+    the form information is added as a new entry.
+
+    Parameters
+    ----------
+    form : flask_wtf.FlaskForm
+        The form being used to provide the data being saved.
+    transaction_id : int
+        The ID of the transaction to be saved. If provided, the
+        named transaction will be updated in the database. Otherwise, if
+        the transaction ID is `None`, a new transaction will be added.
+
+    Returns
+    -------
+    entry : sqlite3.Row
+        An entry corresponding to a transaction in the database.
+
+    Raises
+    ------
+    wtfforms.validators.ValidationError
+        Raised when the form does not validate properly.
+    """
+    if form.validate():
+        transaction_db = BankTransactionHandler()
+        transaction_data = form.transaction_data
+        transfer_data = form.transfer_data
+        if transaction_id:
+            # Update the database with the updated transaction
+            transaction = transaction_db.update_entry(transaction_id,
+                                                      transaction_data)
+        else:
+            # Insert the new transaction into the database
+            if transfer_data:
+                # Update the mappings with the internal transaction information
+                internal_transaction_id = add_internal_transaction()
+                field = 'internal_transaction_id'
+                transfer_data[field] = internal_transaction_id
+                transaction_data[field] = internal_transaction_id
+                # Add the transfer to the database
+                transfer = transaction_db.add_entry(transfer_data)
+            transaction = transaction_db.add_entry(transaction_data)
+        return transaction
+    else:
+        # Show an error to the user and print the errors for the admin
+        flash(form_err_msg)
+        print(form.errors)
+        raise ValidationError('The form did not validate properly.')