+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 
 ## 1.1.0 (in progress)
 - Added banking interface
+- Added interface for displaying linked transactions (including between bank transactions and credit transactions)
 - Grayed out pending transactions
 - Fixed bug where transaction tags appeared in front of header bar
 - Improved modularity of autocomplete JavaScript